config: move fcos packages into fcos subdirectory
